# swagger.yaml
openapi: 3.0.3
info:
  title: th-amaro
  version: 0.2.4
  description: Backend challenge
  contact:
    email: juanfarberjob@gmail.com
servers:
  - url: "http://localhost:9999"
    description: Local API
tags:
  - name: General
    description: Test to see if the API is up
  - name: Products
    description: Products insertions and retrievals

paths:
  /hello:
    get:
      tags:
        - General
      summary: Test to see if the API is up
      description: Test to see if the API is up
      responses:
        "200":
          description: Succesful operation.

  /login:
    post:
      tags:
        - General
      summary: Grants access token given username and password.
      description: Grants access token given username and password.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  description: Username
                  type: string
                  example: juan
                password:
                  description: Password
                  type: string
                  example: juan
              required:
                - username
                - password
      responses:
        "200":
          description: Succesful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: Succesful login.
                  data:
                    description: Products inserted succesfully.
                    type: object
                    example:
                      {
                        "access_token": "v2.local.bY_PBMJKLG8m6R9MRao-ufjkRxfE66gFao5yKRRaBNoWUE_xDFCcBaw-wC3Nk_m4bOmhypXRWn7lZnMmyvCEKSL_TuYvVtGc16Ns7ABq-O9_YyxDvFokaRlHPS4zT6vE77uf4iDK2mbN9emfiTtxKtbL4C3ZtAVbyJC7VMG9InwG_huNuSQndjODVG0DL2Gy9ESc0NFOo8JHExzBxvZ9GNs7TMXQiFM3S15opyDyxm7f7sIL7mYRUiXphwh6ZpHSOABylcj2bS4UGT2J.bnVsbA",
                        "created_at": "2022-10-12T21:55:38.039198356-03:00",
                        "expires_at": "2022-10-12T22:10:38.039201638-03:00",
                      }
        "401":
          description: Unauthorized. Wrong username and/or password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "ERROR. Invalid credentials."
        "500":
          description: Internal server error. Couldn't create access token or couldn't bind parameters properly.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "ERROR. Please, call the administrator."

  /products/batch-insert:
    post:
      tags:
        - Products
      summary: Batch insertion of products and tags.
      description: Batch insertion of products and tags.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                products:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        description: Product ID
                        type: integer
                        example: 8371
                      name:
                        description: Product name
                        type: string
                        example: VESTIDO TRICOT CHEVRON
                      tags:
                        description: Product tags
                        type: array
                        example: ["balada", "neutro", "delicado", "festa"]
              required:
                - products
      responses:
        "200":
          description: Succesful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: Succesful inserts.
                  data:
                    description: Products inserted succesfully.
                    type: array
                    items:
                      type: string
                      example: VESTIDO TRICOT CHEVRON
        "207":
          description: Multi status. Some failed, but could finish gracefully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "ERROR. Failed:"
                  data:
                    description: Products inserted succesfully before the error occurred.
                    type: array
                    items:
                      type: object
                      properties:
                        id_product:
                          type: number
                          example: 8371
                        tags:
                          description: Product tags
                          type: array
                          example: ["balada", "neutro", "delicado", "festa"]
                        error:
                          type: object
                          example:
                            {
                              "Number": 1644,
                              "Message": "id_product already exists.",
                            }
        "500":
          description: Internal server error. Some succeded, but could not finish gracefully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "ERROR. Succesful inserts:"
                  data:
                    description: Products inserted succesfully before the error occurred.
                    type: array
                    items:
                      type: string
                      example: VESTIDO TRICOT CHEVRON
      security:
        - token: []

  /products/search:
    get:
      tags:
        - Products
      summary: Searchs products by id, name and/or tag.
      description: Searchs a product by id, name and/or tag.
      parameters:
        - in: query
          name: id
          description: Product ID
          schema:
            type: integer
          example: 8371
        - in: query
          name: name
          description: Product name
          schema:
            type: string
          example: "VESTIDO TRICOT CHEVRON"
        - in: query
          name: tag
          description: Product tag
          schema:
            type: string
          example: "neutro"
      responses:
        "200":
          description: Succesful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "Succesful search."
                  data:
                    description: Products inserted succesfully.
                    type: object
                    properties:
                      row_count:
                        description: Rows returned.
                        type: number
                        example: 1
                      page_size:
                        description: Rows limit returned by page.
                        type: number
                        example: 1000
                      page_number:
                        description: N° of page offset.
                        type: number
                        example: 1
                      query_result:
                        type: array
                        items:
                          type: object
                          example:
                            {
                              "id_product": 8371,
                              "product": "VESTIDO TRICOT CHEVRON",
                            }
        "400":
          description: Bad Request. Wrong ID format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "Wrong ID format."
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "ERROR."
      security:
        - token: []

components:
  securitySchemes:
    token:
      type: http
      scheme: bearer
      description: Token del usuario